medical_diagnosis/requirements.txt
streamlit
tensorflow
opencv-python
pydicom
numpy
scikit-learn
matplotlib
pillow
